🔧 教育模块v2→v1替换 + da_yun_start_ages: list[int]→float + 测试传参修复

+++ engine/marriage.py
@@ -34,7 +34,7 @@ def analyze_marriage(
     xi_yong: list[str],
     da_yun_gans: list[str],
     da_yun_zhis: list[str],
-    da_yun_start_ages: list[int],
+    da_yun_start_ages: list[float],
 ) -> dict:
     """
     婚姻完整分析
